requirements.txt
docx2txt==0.8
matplotlib==3.4.3
nltk==3.7
numpy==1.19.5
pandas==1.2.4
Pillow==9.1.0
scikit_learn==1.0.2
tensorflow==2.6.0
textdistance==4.2.2
